+++ src/bika/lims/__init__.py
@@ -155,6 +155,7 @@ def initialize(context):
     from bika.lims.controlpanel.bika_worksheettemplates import WorksheetTemplates  # noqa
 
     from bika.lims import permissions
+    from senaite.core import permissions as core_permissions
 
     content_types, constructors, ftis = process_types(
         listTypes(PROJECTNAME), PROJECTNAME)
@@ -165,7 +166,11 @@ def initialize(context):
     for atype, constructor in allTypes:
         kind = "%s: Add %s" % (PROJECTNAME, atype.portal_type)
         perm_name = "Add{}".format(atype.portal_type)
-        perm = getattr(permissions, perm_name, AddPortalContent)
+        # check first if the permission is set in senaite.core
+        perm = getattr(core_permissions, perm_name, None)
+        if perm is None:
+            # check bika.lims.permissions or use fallback permission
+            perm = getattr(permissions, perm_name, AddPortalContent)
         ContentInit(kind,
                     content_types=(atype,),
                     permission=perm,