+++ src/bika/lims/browser/worksheet/views/add_analyses.py
@@ -25,8 +25,6 @@ from bika.lims import bikaMessageFactory as _
 from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.browser.worksheet.tools import showRejectionMessage
 from bika.lims.config import PRIORITIES
-from bika.lims.permissions import EditWorksheet
-from bika.lims.permissions import ManageWorksheets
 from bika.lims.utils import get_image
 from bika.lims.utils import t
 from bika.lims.vocabularies import CatalogVocabulary
@@ -35,6 +33,8 @@ from plone.memoize import view
 from plone.protect import CheckAuthenticator
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from senaite.core.catalog import ANALYSIS_CATALOG
+from senaite.core.permissions.worksheet import can_edit_worksheet
+from senaite.core.permissions.worksheet import can_manage_worksheets
 
 
 class AddAnalysesView(BikaListingView):
@@ -178,15 +178,13 @@ class AddAnalysesView(BikaListingView):
     def is_edit_allowed(self):
         """Check if edit is allowed
         """
-        checkPermission = self.context.portal_membership.checkPermission
-        return checkPermission(EditWorksheet, self.context)
+        return can_edit_worksheet(self.context)
 
     @view.memoize
     def is_manage_allowed(self):
         """Check if manage is allowed
         """
-        checkPermission = self.context.portal_membership.checkPermission
-        return checkPermission(ManageWorksheets, self.context)
+        return can_manage_worksheets(self.context)
 
     def add_status_message(self, message, level="info"):
         """Set a portal status message