+++ src/senaite/core/browser/viewlets/attachments.py
@@ -18,14 +18,14 @@
 # Copyright 2018-2021 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
-from bika.lims import api
 from bika.lims import FieldEditAnalysisResult
-from bika.lims import WorksheetAddAttachment
+from bika.lims import api
 from bika.lims.api.security import check_permission
 from bika.lims.interfaces import IReferenceAnalysis
 from plone.app.layout.viewlets.common import ViewletBase
 from plone.memoize import view
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
+from senaite.core.permissions.worksheet import can_add_worksheet_attachment
 
 
 class AttachmentsViewlet(ViewletBase):
@@ -79,7 +79,7 @@ class WorksheetAttachmentsViewlet(AttachmentsViewlet):
         # XXX: Hack to show the viewlet only on the WS manage_results view
         if not self.request.getURL().endswith("manage_results"):
             return False
-        return check_permission(WorksheetAddAttachment, self.context)
+        return can_add_worksheet_attachment(self.context)
 
     @view.memoize
     def get_services(self):