+++ CHANGES.rst
@@ -4,6 +4,15 @@ Changelog
 2.2.0 (unreleased)
 ------------------
 
+- #1970 Better error messages in sample add form
+- #1960 AddressField and AddressWidget with React component for DX types
+- #1968 Fix default roles for client field in samples
+- #1962 Allow to create worksheet from samples
+- #1966 Fix to set analysis results in batchbooks
+- #1965 Disallow client users to create sample partitions
+- #1964 Fix add samples URL in batch context for Clients
+- #1963 Hide custom set-remarks transition when global remarks are disabled in setup
+- #1961 Added `geo` api  that relies on `pycountry` for retrieval of countries
 - #1911 Converted Container to Dexterity Contents
 - #1931 Removed archetypes.schemaextender from senaite.core
 - #1953 Allow custom id formatting regardless of portal type